build: Ensure schema is up-to-date on build
