+++ .github/workflows/verify-schema.yml
@@ -1,0 +1,19 @@
+name: verify schema
+on:
+  pull_request:
+jobs:
+  build:
+    name: verifying schema
+    runs-on: ubuntu-latest
+
+    steps:
+    - uses: actions/checkout@v3
+
+    - uses: actions/setup-go@v3
+      with:
+        go-version-file: go.mod
+        cache: true
+        cache-dependency-path: go.sum
+
+    - name: Run full tests
+      run: go run ./cmd/schema verify
